refactor: move internal setupkeys operations under operations/
