+++ internal/setupkeys/operations/pub_key_transfer.go
@@ -1,4 +1,4 @@
-package pubkeytransfer
+package operations
 
 import (
 	"context"
