+++ internal/setupkeys/operations/ssh_key_gen.go
@@ -1,4 +1,4 @@
-package sshkeygen
+package operations
 
 import (
 	"fmt"
