+++ internal/setupkeys/operations/ssh_key_gen_test.go
@@ -1,4 +1,4 @@
-package sshkeygen_test
+package operations_test
 
 import (
 	"bytes"
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/arm/topo/internal/setupkeys/sshkeygen"
+	"github.com/arm/topo/internal/setupkeys/operations"
 	"github.com/arm/topo/internal/ssh"
 	"github.com/arm/topo/internal/testutil"
 	"github.com/stretchr/testify/require"
@@ -15,12 +15,12 @@ import (
 
 func TestSSHKeyGenRun(t *testing.T) {
 	keyPath := filepath.Join(t.TempDir(), "keys", "id_ed25519_test")
-	opts := sshkeygen.SSHKeyGenOptions{
+	opts := operations.SSHKeyGenOptions{
 		WithMockKeyGen: func(keyType, keyPath, targetHost string) *exec.Cmd {
 			return testutil.CmdWithOutput("ssh-keygen invoked", 0)
 		},
 	}
-	op := sshkeygen.NewSSHKeyGen("Generate SSH key pair", ssh.NewDestination("user@example.com"), "ed25519", keyPath, opts)
+	op := operations.NewSSHKeyGen("Generate SSH key pair", ssh.NewDestination("user@example.com"), "ed25519", keyPath, opts)
 	var buf bytes.Buffer
 	require.NoError(t, op.Run(&buf))
 	require.Contains(t, buf.String(), "ssh-keygen invoked")
