+++ internal/setupkeys/setup_keys.go
@@ -7,8 +7,7 @@ import (
 
 	"github.com/arm/topo/internal/operation"
 	"github.com/arm/topo/internal/runner"
-	"github.com/arm/topo/internal/setupkeys/pubkeytransfer"
-	"github.com/arm/topo/internal/setupkeys/sshkeygen"
+	"github.com/arm/topo/internal/setupkeys/operations"
 	"github.com/arm/topo/internal/ssh"
 )
 
@@ -22,8 +21,8 @@ const (
 func NewKeySetup(dest ssh.Destination, privKeyPath string, keyType KeyType) (operation.Sequence, error) {
 	sshRunner := runner.NewSSH(dest, runner.SSHOptions{})
 	ops := []operation.Operation{
-		sshkeygen.NewSSHKeyGen("Generate SSH key pair for target", dest, string(keyType), privKeyPath, sshkeygen.SSHKeyGenOptions{}),
-		pubkeytransfer.NewPubKeyTransfer(privKeyPath, sshRunner),
+		operations.NewSSHKeyGen("Generate SSH key pair for target", dest, string(keyType), privKeyPath, operations.SSHKeyGenOptions{}),
+		operations.NewPubKeyTransfer(privKeyPath, sshRunner),
 	}
 	return operation.NewSequence(ops...), nil
 }
